# shared/schemas/api.openapi.json
{
  "components": {
    "schemas": {
...
      "Db2Config": {
        "properties": {
          "default_bind": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Default Bind",
            "type": "object"
          },
          "plan_collection": {
            "title": "Plan Collection",
            "type": "string"
          },
          "subsystem_id": {
            "title": "Subsystem Id",
            "type": "string"
          }
        },
        "required": [
          "subsystem_id",
          "plan_collection"
        ],
        "title": "Db2Config",
        "type": "object"
      },
      "DiffField": {
        "properties": {
          "a": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "A"
          },
          "b": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "B"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "kind"
        ],
        "title": "DiffField",
        "type": "object"
      },
...
      "Envelope_RegionDiffResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegionDiffResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[RegionDiffResponse]",
        "type": "object"
      },
      "Envelope_RegionForkResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegionForkResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[RegionForkResponse]",
        "type": "object"
      },
      "Envelope_RegionListResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegionListResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[RegionListResponse]",
        "type": "object"
      },
      "Envelope_RegionProfile_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegionProfile"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[RegionProfile]",
        "type": "object"
      },
      "Envelope_RegionUpsertResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegionUpsertResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[RegionUpsertResponse]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JesConfig": {
        "properties": {
          "class": {
            "title": "Class",
            "type": "string"
          },
          "sysout_class": {
            "title": "Sysout Class",
            "type": "string"
          }
        },
        "required": [
          "class",
          "sysout_class"
        ],
        "title": "JesConfig",
        "type": "object"
      },
...
      "RegionDiffResponse": {
        "properties": {
          "a": {
            "title": "A",
            "type": "string"
          },
          "b": {
            "title": "B",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/DiffField"
            },
            "title": "Fields",
            "type": "array"
          }
        },
        "required": [
          "a",
          "b",
          "fields"
        ],
        "title": "RegionDiffResponse",
        "type": "object"
      },
      "RegionForkRequest": {
        "properties": {
          "overrides": {
            "additionalProperties": true,
            "title": "Overrides",
            "type": "object"
          },
          "reason": {
            "maxLength": 400,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "overrides",
          "reason"
        ],
        "title": "RegionForkRequest",
        "type": "object"
      },
      "RegionForkResponse": {
        "properties": {
          "audit_event_id": {
            "title": "Audit Event Id",
            "type": "string"
          },
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          }
        },
        "required": [
          "region",
          "job_name",
          "audit_event_id"
        ],
        "title": "RegionForkResponse",
        "type": "object"
      },
      "RegionListItem": {
        "properties": {
          "hlq": {
            "title": "Hlq",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tier": {
            "$ref": "#/components/schemas/RegionTier"
          }
        },
        "required": [
          "name",
          "tier",
          "hlq"
        ],
        "title": "RegionListItem",
        "type": "object"
      },
      "RegionListResponse": {
        "properties": {
          "regions": {
            "items": {
              "$ref": "#/components/schemas/RegionListItem"
            },
            "title": "Regions",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "regions",
          "total"
        ],
        "title": "RegionListResponse",
        "type": "object"
      },
      "RegionProfile": {
        "additionalProperties": true,
        "description": "Full region document \u2014 keeps shape in sync with docs/DATA_MODEL.md \u00a73.",
        "properties": {
          "confidence_weight_overrides": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Confidence Weight Overrides",
            "type": "object"
          },
          "db2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Db2Config"
              },
              {
                "type": "null"
              }
            ]
          },
          "gdg_retention": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gdg Retention"
          },
          "hlq": {
            "title": "Hlq",
            "type": "string"
          },
          "jes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JesConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "protected_resources": {
            "items": {
              "type": "string"
            },
            "title": "Protected Resources",
            "type": "array"
          },
          "racf_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Racf Group"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewConfig"
          },
          "scheduler_queue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduler Queue"
          },
          "tier": {
            "$ref": "#/components/schemas/RegionTier"
          },
          "volser_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volser Pattern"
          }
        },
        "required": [
          "name",
          "tier",
          "hlq"
        ],
        "title": "RegionProfile",
        "type": "object"
      },
      "RegionTier": {
        "enum": [
          "development",
          "integration",
          "qa",
          "uat",
          "production"
        ],
        "title": "RegionTier",
        "type": "string"
      },
      "RegionUpsertRequest": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/RegionProfile"
          },
          "reason": {
            "maxLength": 400,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "profile",
          "reason"
        ],
        "title": "RegionUpsertRequest",
        "type": "object"
      },
      "RegionUpsertResponse": {
        "properties": {
          "audit_event_id": {
            "title": "Audit Event Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "review_required": {
            "title": "Review Required",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "audit_event_id",
          "review_required"
        ],
        "title": "RegionUpsertResponse",
        "type": "object"
      },
      "ReviewConfig": {
        "properties": {
          "allowed_reviewers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Allowed Reviewers",
            "type": "object"
          },
          "auto_approve_threshold": {
            "default": 90,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Auto Approve Threshold",
            "type": "integer"
          }
        },
        "title": "ReviewConfig",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI control plane for z/OS modernization. See docs/API.md.",
    "title": "Helios Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/regions": {
      "get": {
        "operationId": "list_regions_api_regions_get",
        "parameters": [
          {
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RegionTier"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RegionListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Regions",
        "tags": [
          "regions"
        ]
      }
    },
    "/api/regions/{a}/diff/{b}": {
      "get": {
        "operationId": "diff_regions_api_regions__a__diff__b__get",
        "parameters": [
          {
            "in": "path",
            "name": "a",
            "required": true,
            "schema": {
              "title": "A",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "b",
            "required": true,
            "schema": {
              "title": "B",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RegionDiffResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff Regions",
        "tags": [
          "regions"
        ]
      }
    },
    "/api/regions/{name}": {
      "get": {
        "operationId": "get_region_api_regions__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RegionProfile_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Region",
        "tags": [
          "regions"
        ]
      },
      "post": {
        "operationId": "upsert_region_api_regions__name__post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RegionUpsertResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Region",
        "tags": [
          "regions"
        ]
      }
    },
    "/api/regions/{name}/forks/{job_name}": {
      "post": {
        "operationId": "fork_region_api_regions__name__forks__job_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionForkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RegionForkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fork Region",
        "tags": [
          "regions"
        ]
      }
    },
...
  }
}